+++ R/pandoc.R
@@ -308,20 +308,12 @@ pandoc_include_args <- function(in_header = NULL,
 
 #' @rdname pandoc_args
 #' @export
-pandoc_highlight_args <- function(highlight,
-                                  default = "tango") {
-
-  args <- c()
-
-  if (is.null(highlight))
-    args <- c(args, "--no-highlight")
-  else {
-    if (identical(highlight, "default"))
-      highlight <- default
-    args <- c(args, "--highlight-style", highlight)
+pandoc_highlight_args <- function(highlight, default = "tango") {
+  if (is.null(highlight)) {
+    if (pandoc_available("3.8")) "--syntax-highlighting=none" else "--no-highlight"
+  } else {
+    c("--highlight-style", if (identical(highlight, "default")) default else highlight)
   }
-
-  args
 }
 
 #' @rdname pandoc_args